feat: added some celery + rabbitmq example
