--- .github/workflows/run-tests.yml
@@ -1,87 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-name: Run tests
-
-# Controls when the workflow will run
-on:
-  # Triggers the workflow on push or pull request events
-  [ push ]
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-
-  # This workflow contains a single job called "style"
-  run-tests:
-
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    services:
-      rabbitmq:
-        image: rabbitmq
-        env:
-          RABBITMQ_DEFAULT_USER: megan
-          RABBITMQ_DEFAULT_PASS: megan123
-          RABBITMQ_DEFAULT_VHOST: megan_vhost
-        ports:
-          - 5672
-      celery-1:
-        image: celery
-        env:
-          BROKER_URL: amqp://megan:megan123@localhost/megan_vhost
-          CELERY_INCLUDE: homework.tasks
-#          BROKER_TRANSPORT:
-
-    # A strategy creates a build matrix for your jobs
-    strategy:
-
-      # You can define a matrix of different job configurations
-      matrix:
-
-        # Each option you define in the matrix has a key and value
-        python-version: [ 3.11 ]
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - name: Set up Git repository
-        uses: actions/checkout@v2
-
-      # Setup Python with version from matrix
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v3
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      # Install requirements
-      - name: Install requirements
-
-        # Runs command-line programs using the operating system's shell
-        run: |
-          python -m pip install --upgrade pip wheel setuptools
-          python -m pip install -r requirements.txt
-          python -m pip list
-
-      # Install pre-commit from .pre-commit-config.yaml
-      - name: Install pre-commit
-        run: |
-          pre-commit install
-
-      # Run pre-commit on all the files in the repo
-      - name: Run pre-commit
-        run: |
-          pre-commit run --all-files --color always --verbose --show-diff-on-failure
-
-#      - name: Run celery worker
-#        run: |
-#          celery -A homework worker --loglevel=info -n work1
-#          celery -A homework worker --loglevel=info -n work2
-#          celery -A homework worker --loglevel=info -n work3
-#        env:
-#          RABBITMQ_HOST: amqp://megan:megan123@localhost/
-#          RABBITMQ_PORT: ${{ job.services.rabbitmq.ports['5672'] }
-
-      # Run tests
-      - name: Run tests
-        run: |
-          pytest tests/
