--- DockerfileCelery
@@ -1,12 +1,0 @@
-FROM python:alpine
-
-COPY requirements.txt requirements.txt
-
-RUN /usr/local/bin/python -m pip install --upgrade pip
-
-RUN pip install -r requirements.txt
-
-COPY ./homework /homework
-WORKDIR homework
-
-ENTRYPOINT celery -A tasks worker --loglevel=info
