--- DockerfileFastapi
@@ -1,13 +1,0 @@
-FROM python:alpine
-
-COPY requirements.txt requirements.txt
-
-RUN /usr/local/bin/python -m pip install --upgrade pip
-
-RUN pip install -r requirements.txt
-
-COPY ./app /fastapi
-WORKDIR fastapi
-
-EXPOSE 80
-CMD ["python", "main.py"]
