+++ docker-compose.yaml
@@ -9,40 +9,33 @@ services:
     environment:
       RABBITMQ_DEFAULT_USER: megan
       RABBITMQ_DEFAULT_PASS: megan123
-    restart: unless-stopped
+      RABBITMQ_DEFAULT_VHOST: megan_vhost
+    restart: always
     networks:
-      - backend-net
+      - net
 
   fastapi:
-    build:
-      context: .
-      dockerfile: DockerfileFastapi
-    image: fastapi
-    container_name: app
+    build: .
+    command: poetry run uvicorn homework.app.main:app --reload
     ports:
       - "80:80"
     environment:
-      RMQ_URL: 'amqp://megan:megan123@rabbitmq:5672'
+      RMQ_URL: 'amqp://megan:megan123@rabbitmq:5672/megan_vhost'
     depends_on:
       - rabbitmq
-    restart: unless-stopped
     networks:
-      - backend-net
+      - net
 
-  celery:
-    build:
-      context: .
-      dockerfile: DockerfileCelery
-    image: celery
-    container_name: celery
+  work:
+    build: .
+    command: poetry run celery -A homework.tasks worker --loglevel=info -n work
     environment:
-      RMQ_URL: 'amqp://megan:megan123@rabbitmq:5672'
+      RMQ_URL: 'amqp://megan:megan123@rabbitmq:5672/megan_vhost'
     depends_on:
       - fastapi
-    restart: unless-stopped
     networks:
-      - backend-net
+      - net
 
 networks:
-  backend-net:
+  net:
     driver: bridge
