+++ homework/models/enums.py
